+++ TR5Main/Specific/trmath.h
@@ -33,7 +33,7 @@ BoundingOrientedBox TO_DX_BBOX(PHD_3DPOS pos, BOUNDING_BOX* box);
 const float lerp(float v0, float v1, float t);
 const Vector3 getRandomVector();
 const Vector3 getRandomVectorInCone(const Vector3& direction, const float angleDegrees);
-int mGetAngle(int x1, int y1, int x2, int y2);
+float mGetAngle(int x1, int y1, int x2, int y2);
 void phd_GetVectorAngles(int x, int y, int z, float* angles);
 void phd_RotBoundingBoxNoPersp(PHD_3DPOS* pos, BOUNDING_BOX* bounds, BOUNDING_BOX* tbounds);
 int phd_Distance(PHD_3DPOS* first, PHD_3DPOS* second);
